Changes following John's review
