+++ test/testdata/fixtures/multi-dc/k8ssandra.yaml
@@ -25,7 +25,7 @@ spec:
         nodeAffinityLabels:
           "topology.kubernetes.io/zone": rack3
     config:
-      cassandra-yaml:
+      cassandraYaml:
         auto_snapshot: false
         memtable_flush_writers: 1
         commitlog_segment_size_in_mb: 2
