+++ test/testdata/fixtures/multi-stargate/cassdc.yaml
@@ -5,7 +5,7 @@ metadata:
 spec:
   clusterName: test
   config:
-    cassandra-yaml:
+    cassandraYaml:
       auto_snapshot: false
       commitlog_segment_size_in_mb: 2
       compaction_throughput_mb_per_sec: 0
